Improved performance of Android router (#192)
* mitigate race condition at router startup

When there are a lot of network announcements to load, channels
could start before the router had time to subscribe to
`ChannelStateChanged` and missed events. This resulted in local
channels not being detected by the router, which in turn led to 
`RouteNotFound` errors.

* optimized router for android

When initializing the router, the stored announcements are directly
loaded in memory and don't go to the 'normal' process involving signature
verification, events and all.

We also do not broadcast announcements anymore.