added more checks on `open`/`accept` messages
We check that channel parameters are compliant with [BOLT
2](https://github.com/lightningnetwork/lightning-rfc/blob/master/02-peer-protocol.md)
(this fixes #236).

We make sure that the counterparty chooses a decent `dust_limit`
because we want them to be able to publish their commitment, e.g. in a
data loss scenario. We also make sure that our configurable `dust_limit` isn't too low (this
fixes #234).

Also fixed our min `dust_limit` (542->546).