Disabled JSON-RPC API by default. BASIC auth is required when enabled
The JSON RPC API should be disabled by default because it enables possible
attacks on the node. The user must be aware of this and has to enable it himself.

The API also requires a user/password to be set in `eclair.conf`. If the API is
enabled and the user/password are not set, the node can not be started.

CORS is disabled.