NodeURI can now handle ipv6 addresses
Using HostAndPort from guava to parse a LN address string and extract
the host and port.
Parser defaults to port 9735 if the port could not be found.

Updated the node uri regex to handle the IPV6 special characters.

IPV6 addresses are surrounded by brackets so that a parser can
differentiate between the host and the actual port of the node.