Channel capacity is saved in network DB
The router also saves the channel capacity along with the tx id when
discovering a new channel. `ChannelDiscovered` now contains the
capacity.

The gui adds a column in the 'all channels' table to display the
the capacity.

Added a compatibility check for the network DB in startup. This check is
separated from the node DB check because a network DB check failure is
less severe and the network DB file can be safely removed with no impact
on the node.