use a `DirectedWeightedPseudograph`
Because we want a directed graph with multiple weighted edges between two
vertices, and loops allowed.

See discussion in [1] for more details.

Also, followed @btcontract's suggestion of using a
`BidirectionalDijkstraShortestPath` instead of `DijkstraShortestPath`.

In `findRoute`, if we have specific updates to add/remove, we don't mutate
the main graph and create a copy instead.

[1]
http://jgrapht-users.107614.n3.nabble.com/Difference-between-a-directed-multigraph-and-a-directed-pseudograph-td4024788.html