separate close types `MUTUAL`/`FORCE`
A new close type (`MUTUAL`/`FORCE`) has been added to `CMD_CLOSE`:
- If type is `MUTUAL`, then the command will succeed only if the current
channel state is in NORMAL, or if the channel hasn't yet been created.
- If type is `FORCE`, then the channel will publish its current local
commitment (or its best signed closing tx if it has one).

Using `FORCE` close type is more expensive and it incurs a delay before
funds are spendable, but this can be useful in some situations, for
example when the counterparty isn't responding anymore.

A new `forceclose` API method has been added.