resolve htlcs using the amount too
We may have htlcs with identical `payment_hash` + `cltv_expiry`, but
different amounts.

In that case we need to use the `amount` to differentiate and choose the
correct commitment output for each htlc.

In other cases, when there is no risk of confusion it is ok to only rely
on `pubkeyScript`.