relay htlcs to channels with the highest balance
In order to reduce unnecessary back-and-forth in case an outgoing
channel doesn't have enough capacity but another one has, the relayer
can now forward a payment to a different channel that the one specified
in the onion (to the same node of course).

As a side effect, if this preferred channel returns an error, then the
sender will have an error for a different channel than the one it
selected. Note that eclair already supports that (see PaymentLifecycle).