make router handle updates from failure messages
This is a regression caused by 9f708acf04f11551ec6a6c9147f1c9fcfba737e7,
which made the `Peer` class encapsulates network announcements inside
`PeerRoutingMessage`, in order to preserve the origin of the messages.

But when we get channel updates as part of failure messages when making
payments, they weren't encapsulated and were ignored by the router.

Re-enabled a non-regression test in `IntegrationSpec` which will prevent
this from happening in the future.