Add `pendingPaymentDb` to track payments which take too many blocks
First, this can be used to evaluate whether a payment should be sent to a given payee. We can never be 100% certain since payment may be delayed by final payee as well as by any routing node, but if the same payee has many delayed payments for some period of time that could be telling.

Second, a specific case can be catched where we are certain about what's going on: when payment is delayed beyond a first hop cltv expiry we do know our peer should have failed it already but it did not. This can be used to evaluate whether it's worthy to send more payments through this peer or open new channels with it.